+++ lib/axlsx.rb
@@ -10,6 +10,7 @@ require 'axlsx/util/accessors.rb'
 require 'axlsx/util/serialized_attributes'
 require 'axlsx/util/options_parser'
 require 'axlsx/util/mime_type_utils'
+require 'axlsx/util/zip_command'
 
 require 'axlsx/stylesheet/styles.rb'